Add files via upload
Still on the topic of Data cleaning with Excel Power Query, I did a little bit of web scrapping where I went on Wikipedia and searched for the population of the United States of America and uploaded it into my editor and transformed the data, removed some columns that I did not need, renamed columns, unpivoted the columns, extracted the year. Please check it out and kindly let me know what you think.